requirements.txt
requests
beautifulsoup4
lxml
openai
openrouter
pytest==7.4.0
streamlit
cryptography==42.0.8
gigachat
transformers 
torch --index-url https://download.pytorch.org/whl/cpu
torchvision 
pillow 
beautifulsoup4==4.12.3
mnemonic
# CTF / API
pyyaml
fastapi
uvicorn[standard]
aiohttp
# PDF reports (Iceo96/Ai-agent-WebSecurity style)
reportlab
